+++ src/components/pages/DevCornerOneDashboard/index.ts
@@ -1,1 +1,1 @@
-export * from './DevCornerOneDashboard';
+export { DevCornerOneDashboard } from './DevCornerOneDashboard';